backend/requirements.txt
langchain
langchain-community
python-dotenv
pydantic
fastapi
pypdf
google-generativeai
langchain_google_genai
uvicorn[standard]
pipdeptree
Markdown
IPython
pdfplumber